+++ HsBlog_BackEnd/src/main/java/com/hs/hsblog_backend/repository/FriendItemMapper.java
@@ -16,7 +16,7 @@ public interface FriendItemMapper {
 
     void updateFriendItem(FriendItem friendItem);
 
-    void deleteFriendItemById(Integer id);
+    void deleteFriendItemById(Long id);
 }
